created morph model: Meta to limit meta tables
