+++ app/Http/Controllers/MetaController.php
@@ -1,0 +1,85 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Meta;
+use Illuminate\Http\Request;
+
+class MetaController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Meta  $meta
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Meta $meta)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\Meta  $meta
+     * @return \Illuminate\Http\Response
+     */
+    public function edit(Meta $meta)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Meta  $meta
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, Meta $meta)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Meta  $meta
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy(Meta $meta)
+    {
+        //
+    }
+}
