--- app/UserMeta.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class UserMeta extends Model
-{
-  protected $fillable = ['user_id', 'name', 'value'];
-
-  public function user(){
-    return $this->belongsTo(User::class);
-  }
-}
